Add the preset artifact type
