+++ packages/lms-shared-types/src/PluginManifest.ts
@@ -1,4 +1,5 @@
 import { z } from "zod";
+import { artifactManifestBaseSchema, type ArtifactManifestBase } from "./ArtifactManifestBase.js";
 
 /**
  * @public
@@ -6,25 +7,15 @@ import { z } from "zod";
 export type PluginRunnerType = "ecmascript";
 export const pluginRunnerTypeSchema = z.enum(["ecmascript"]);
 
-export const kebabCaseRegex = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
-export const kebabCaseSchema = z.string().regex(kebabCaseRegex);
-
 /**
  * @public
  */
-export interface PluginManifest {
+export interface PluginManifest extends ArtifactManifestBase {
   type: "plugin";
   runner: PluginRunnerType;
-  owner: string;
-  name: string;
-  description: string;
-  revision?: number;
 }
 export const pluginManifestSchema = z.object({
   type: z.literal("plugin"),
   runner: pluginRunnerTypeSchema,
-  owner: kebabCaseSchema,
-  name: kebabCaseSchema,
-  description: z.string(),
-  revision: z.number().int().optional(),
+  ...artifactManifestBaseSchema.shape,
 });
